+++ test/audits/forms/formcalc.test.js
@@ -31,6 +31,7 @@ describe('Form Calc functions', () => {
         pageview: { total: 8670, desktop: 4670, mobile: 4000 },
         url: 'https://www.surest.com/info/win',
         trafficacquisition: {},
+        formsource: '.myform',
       },
       {
         formengagement: { total: 300, desktop: 0, mobile: 300 },
@@ -39,6 +40,7 @@ describe('Form Calc functions', () => {
         pageview: { total: 8670, desktop: 4670, mobile: 4000 },
         url: 'https://www.surest.com/newsletter',
         trafficacquisition: {},
+        formsource: '',
       },
     ]);
   });
@@ -48,15 +50,21 @@ describe('Form Calc functions', () => {
     expect(result).to.eql([
       {
         url: 'https://www.surest.com/info/win',
-        pageViews: 8670,
-        formViews: 300,
-        formEngagement: 4300,
+        formengagement: { total: 4300, desktop: 4000, mobile: 300 },
+        formsubmit: { total: 0, desktop: 0, mobile: 0 },
+        formview: { total: 300, desktop: 0, mobile: 300 },
+        pageview: { total: 8670, desktop: 4670, mobile: 4000 },
+        trafficacquisition: {},
+        formsource: '.myform',
       },
       {
         url: 'https://www.surest.com/newsletter',
-        pageViews: 8670,
-        formViews: 300,
-        formEngagement: 300,
+        formengagement: { total: 300, desktop: 0, mobile: 300 },
+        formsubmit: { total: 0, desktop: 0, mobile: 0 },
+        formview: { total: 300, desktop: 0, mobile: 300 },
+        pageview: { total: 8670, desktop: 4670, mobile: 4000 },
+        trafficacquisition: {},
+        formsource: '',
       },
     ]);
   });
@@ -70,6 +78,7 @@ describe('Form Calc functions', () => {
         formview: { total: 300, desktop: 0, mobile: 300 },
         formengagement: { total: 300, desktop: 0, mobile: 300 },
         formsubmit: { total: 0, desktop: 0, mobile: 0 },
+        formsource: '',
         trafficacquisition: {},
         CTA: {
           url: 'https://www.surest.com/about-us',