+++ test/fixtures/forms/high-form-views-low-conversions.js
@@ -335,6 +335,63 @@ const testData = {
       ],
     },
   },
+  lowFormviewsAuditData: {
+    type: 'high-page-views-low-form-nav',
+    siteId: 'site-id',
+    auditId: 'audit-id',
+    auditResult: {
+      formVitals: [
+        {
+          url: 'https://www.surest.com/high-page-low-form-view',
+          formsubmit: {
+            'desktop:windows': 0,
+          },
+          formview: {
+            'desktop:windows': 200,
+          },
+          formengagement: {
+            'desktop:windows': 100,
+          },
+          pageview: {
+            'desktop:windows': 5690,
+            'mobile:ios': 1000,
+          },
+          trafficacquisition: {
+            paid: 2690,
+            maxTimeDelta: 3060,
+            total: 6690,
+            earned: 2000,
+            sources: [],
+            owned: 2000,
+          },
+        },
+        {
+          url: 'https://www.surest.com/existing-opportunity',
+          formsubmit: {
+            'desktop:windows': 0,
+          },
+          formview: {
+            'desktop:windows': 200,
+          },
+          formengagement: {
+            'desktop:windows': 100,
+          },
+          pageview: {
+            'desktop:windows': 5690,
+            'mobile:ios': 1000,
+          },
+          trafficacquisition: {
+            paid: 2690,
+            maxTimeDelta: 3060,
+            total: 6690,
+            earned: 2000,
+            sources: [],
+            owned: 2000,
+          },
+        },
+      ],
+    },
+  },
   auditData3: {
     type: 'high-form-views-low-conversions',
     siteId: 'site-id',
@@ -361,6 +418,7 @@ const testData = {
         },
         {
           url: 'https://www.surest.com/contact-us',
+          formsource: '.mycontact',
           formsubmit: {
             'desktop:windows': 100,
           },
@@ -397,6 +455,7 @@ const testData = {
         },
         {
           url: 'https://www.surest.com/info/win-1',
+          formsource: '.form',
           formsubmit: {
             'desktop:windows': 100,
           },
@@ -762,6 +821,7 @@ const testData = {
       screenshot: '',
       formViews: 3000,
       pageViews: 8670,
+      formsource: '',
       samples: 8670,
       scrapedStatus: false,
     },
@@ -785,6 +845,7 @@ const testData = {
       trackedFormKPIValue: 0.014947683109118086,
       formViews: 6690,
       pageViews: 6690,
+      formsource: '',
       samples: 6690,
       scrapedStatus: false,
       metrics: [
@@ -883,6 +944,7 @@ const testData = {
       trackedFormKPIValue: 0.014947683109118086,
       formViews: 6690,
       pageViews: 6690,
+      formsource: '',
       samples: 6690,
       scrapedStatus: false,
       metrics: [
@@ -990,6 +1052,7 @@ const testData = {
       trackedFormKPIValue: 0.014947683109118086,
       formViews: 6690,
       pageViews: 6690,
+      formsource: '',
       samples: 6690,
       scrapedStatus: true,
       metrics: [
@@ -1171,6 +1234,7 @@ const testData = {
       form: 'https://www.surest.com/contact-us',
       formViews: 3000,
       pageViews: 7690,
+      formsource: '',
       screenshot: '',
       samples: 7690,
       scrapedStatus: false,